refactor🧰: directories and types
