+++ src/components/Card/Card.tsx
@@ -18,11 +18,7 @@ interface IProps {
 }
 
 const Card: FC<IProps> = React.memo(
-  ({
-    card: { count, year, shape, color, size, favorite, name, num },
-    onChange,
-    value,
-  }) => {
+  ({ card: { count, year, shape, color, size, favorite, name, num } }) => {
     const { favoriteCards, setFavoriteCards } = useContext(FavoriteContext);
     const setFavoriteCardsArray = (checked: boolean, number: string) => {
       const cardIndex = parseInt(number) - 1;
