+++ src/components/FavoriteCards/FavoriteCards.module.scss
@@ -1,4 +1,4 @@
-@import "src/style-dev/vars";
+@import "../../style-dev/vars";
 
 .card {
   position: relative;
