+++ src/components/Lights/Lights.module.scss
