+++ src/components/Lights/Lights.tsx
