+++ src/components/Picture/Picture.module.scss
