+++ src/pages/Game.module.scss
